添加snake.cpp，单独放置class Snake；蛇不再可以向后运动

+++ PainterEngine/project/PainterEngine_Application.cpp
@@ -62,7 +62,7 @@ px_bool PX_ApplicationInitialize(PX_Application *pApp,px_int screen_width,px_int
 		keyMap_2.skill_2 = VK_NUMPAD2;
 		keyMap_2.slowdown = VK_NUMPAD0;
 			
-		player[1].Init(2, keyMap_2, MAP_SIZE_X / 1.5, MAP_SIZE_Y / 2, 30, PX_COLOR(255, 157, 208, 136));
+		player[1].Init(2, keyMap_2, (int)(MAP_SIZE_X / 1.5), MAP_SIZE_Y / 2, 30, PX_COLOR(255, 157, 208, 136));
 	}
 
 	InitializeCriticalSection(&g_cs);//初始化临界区
